feat: 수강 등록 시 결제 금액 검증 및 요청/응답 타입 long 통일
- EnrollmentRequestDTO에 paidAmount 필드 추가 (@Positive, @Max, @Schema)
- CancelRequestDTO cancelAmount int → long
- EnrollmentResponseDTO, CancelResponseDTO paidAmount/cancelAmount int → long
- EnrollmentService: 결제 금액과 강의 정가 불일치 시 INVALID_PAYMENT_AMOUNT(400) 반환
- ErrorCode: INVALID_PAYMENT_AMOUNT 추가

Made-with: Cursor

+++ creator-settlement-pjt/src/main/java/com/seohyun/creator_settlement_pjt/dto/CancelResponseDTO.java
@@ -12,8 +12,8 @@ public class CancelResponseDTO {
     private Long saleRecordId;
     private String courseTitle;
     private String studentName;
-    private int paidAmount;
-    private int cancelAmount;
+    private long paidAmount;
+    private long cancelAmount;
     private LocalDateTime canceledAt;
 
 }
